AT LAST! VitIO disk driver did it! Still need to finish driver, but disk IO is working now.

+++ trunk/include/virtio_blk.h
@@ -54,7 +54,7 @@ struct virtio_blk_outhdr
 	/* io priority. */
 	u_int32_t ioprio;
 	/* Sector (ie. 512 byte offset) */
-	u_int32_t sector;
+	u_int64_t sector;
 } __packed;
 
 /* And this is the final byte of the write scatter-gather list. */
